Fix JSmol stuff
This script allows you to either store webpages showing:
- the reactant, saddle point, and product state or
- a movie from reactant to product.
Movies will play in a continuous loop. Any structure can be viewed
from different angles, and zoom levels. Right-clicking on the JSmol
"window" allows you to change display settings as well.

The main disadvantage is that for reasons of cybersecurity the
page needs to be loaded from a webserver. Also, JSmol needs to
installed inside the data directory.

So your directory structure needs to be something like

- states/
  - index.html
  - jmol/
    - jsmol/
  - 0/
    - procdata/
      - reaction_0.xyz
      - reaction_1.xyz
      - reaction_2.xyz

The "jmol" directory can be a soft link to your real Jmol installation.
In this directory structure you need to start the webserver inside the
"states" directory. I.e.

    cd states
    python -m http.server 8000

then in your webbrowser navigate to

    http://localhost:8000/

to display the web page properly. The ironic side effect of course
is that for cybersecurity reasons you now have to make your own data
world accessible on a webserver...